pull request changes: null pointer check, timeout function, error returns, function names, rx nack logic. bug fix: enable master mode on init.
